feat: make instruction handler generic
